refactor: separate helpers
