+++ src/lib/helpers/transformations.ts
@@ -1,0 +1,42 @@
+import type { SelectInputOption } from "@/types/forms";
+
+/** Converts a TypeScript enum into an array of form select options. */
+export const enumToFormSelectOptions = <T extends string | number>(
+  enumObject: Record<string, T>,
+  labels: Record<T, string>,
+): SelectInputOption[] =>
+  Object.entries(enumObject)
+    .filter(([key]) => Number.isNaN(Number(key)))
+    .map(([, value]) => ({
+      value,
+      label: labels[value],
+    }));
+
+export function sanitizeId(id: string | number): string {
+  return String(id).trim().replaceAll(/[^\d]/g, "");
+}
+
+export function removeTrailingSlash(path: string): string {
+  return path.replace(/\/+$/, "");
+}
+
+/** This needs to be used instead of encodeURIComponent for requests made to the backend because it cannot parse '%20' as space, only the plus symbol. */
+export const encodeQueryComponent = (value: string) =>
+  encodeURIComponent(value).replaceAll("%20", "+");
+
+/** Encodes the given parameter values as a string using `encodeQueryComponent`, omitting any keys with null or undefined values. */
+export const encodeQueryParameters = (
+  parameters: Record<string, string | null | undefined>,
+): string => {
+  const pairArray = Object.entries(parameters).reduce<string[]>(
+    (pairs, [key, value]) => {
+      if (value != null) {
+        const encodedPair = `${key}=${encodeQueryComponent(value)}`;
+        pairs.push(encodedPair);
+      }
+      return pairs;
+    },
+    [],
+  );
+  return pairArray.join("&");
+};
