+++ src/lib/helpers/validation.ts
@@ -1,0 +1,11 @@
+import { z } from "zod";
+
+import { FORM_ERROR_MESSAGES } from "@/config/constants";
+
+export const requiredString = () =>
+  z
+    .string({ required_error: FORM_ERROR_MESSAGES.REQUIRED })
+    .trim()
+    .min(1, { message: FORM_ERROR_MESSAGES.NONEMPTY });
+
+export const timestampNumber = () => z.number().int().positive();
